Added base class
